% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{sample_raster}
\alias{sample_raster}
\title{Example RGB GeoTIFF (UTM 35N)}
\format{
A GeoTIFF file in `inst/extdata/`.
}
\description{
A small 3-band (RGB) GeoTIFF used in examples, vignettes, and tests.
}
\details{
The file is shipped with the package at:
`system.file("extdata", "sample_raster.tif", package = "rsegm")`.

Read it with:
`terra::rast(system.file("extdata","sample_raster.tif", package="rsegm", mustWork=TRUE))`.
}
\keyword{datasets}
