% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{sample_raster_path}
\alias{sample_raster_path}
\title{Path to an example raster}
\usage{
sample_raster_path()
}
\value{
A file path to an example GeoTIFF shipped with the package.
}
\description{
Path to an example raster
}
